refactor: 모든 TABS import 제거 후 상수 파일을 참조하도록 변경
- 순환 의존성 해소
